pyproject.toml
[tool.poetry]
name = "test-task-anverali"
version = "0.1.0"
description = ""
authors = ["Kuragin Ivan <curagin.vanya@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
sqlalchemy = "^2.0.29"
alembic = "^1.13.1"
flask = "^3.0.3"
flask-admin = "^1.6.1"
gunicorn = "^22.0.0"
python-dotenv = "^1.0.1"
flask-login = "^0.6.3"
psycopg2-binary = "^2.9.9"
flasgger = "^0.9.7.1"
marshmallow = "^3.21.1"
webargs = "^8.4.0"
requests = "^2.31.0"
sentry-sdk = {extras = ["flask"], version = "^2.0.1"}
pytest = "^8.2.0"
opensearch-py = "^2.5.0"
bcrypt = "^4.1.2"
opensearch-logger = "^1.3.0"

[tool.poetry.group.dev.dependencies]
flake8 = "^7.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
